backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Pillow==10.0.1
bcrypt==4.0.1
